% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDDfill.R
\name{DDDfill}
\alias{DDDfill}
\title{function to fill missing ddd-values from drug purchase data}
\usage{
DDDfill(
  read_path,
  k_mean = 1,
  tol = 10,
  pvmvar = "otpvm",
  vnrovar = "vnr",
  kustvar = "kust",
  dddvar = "ddd",
  restoreOrder = 1
)
}
\arguments{
\item{read_path}{path to data (supported fileforms: .sas7bdat)}

\item{k_mean}{1 (default), >1 compares the closest prior match with (up to) k_mean matches on both sides}

\item{tol}{10 (default), tolerance of rejecting the closest value when comparing to the mean value}

\item{pvmvar}{"otpvm" (default), what is the column id for purchase dates}

\item{vnrovar}{"vnr" (default), what is the column id for purchase code}

\item{kustvar}{"kust" (default), what is the column id for cost of the purchase}

\item{dddvar}{"ddd" (default), what is the column id for ddd}

\item{restoreOrder}{1 (default), with default restores the ordering to original,
with 0 the returned data.table is ordered by pvmvar and then by vnrovar}
}
\value{
ddd filled data.table
}
\description{
Fills the missing ddd-values in the data according to following logic:
by default finds the closest prior vnr-matching purchase (if prior not found, finds closest after).
However, if k_mean given to be different from 1, compares the match with the trimmed mean of (up to) k_mean
matching purchases both sides to the purchase, and uses that if too big of a difference (reduces possible wrong entry effect).
For match, calculates the ddd/kust -ratio, which can be then used to fill the missing ddd-value by:
kust of ddd-missing purchase * ddd/kust -ratio (of closest match or mean of kmean matches around the purchase)
}
\examples{
# example call with alternative values for some of the default valued parameters:
ddd_filled_data.table <- DDDfill("path_to_original_data.sas7bdat", k_mean = 4)
}
